[flink] Introduce new ActionBase
