+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/action/ActionBase.java
@@ -18,148 +18,31 @@
 
 package org.apache.paimon.flink.action;
 
-import org.apache.paimon.CoreOptions;
 import org.apache.paimon.catalog.Catalog;
-import org.apache.paimon.catalog.Identifier;
 import org.apache.paimon.flink.FlinkCatalog;
 import org.apache.paimon.flink.FlinkCatalogFactory;
-import org.apache.paimon.flink.LogicalTypeConversion;
-import org.apache.paimon.flink.sink.FlinkSinkBuilder;
-import org.apache.paimon.flink.utils.TableEnvironmentUtils;
-import org.apache.paimon.operation.Lock;
 import org.apache.paimon.options.CatalogOptions;
 import org.apache.paimon.options.Options;
-import org.apache.paimon.table.FileStoreTable;
-import org.apache.paimon.table.Table;
-import org.apache.paimon.types.DataType;
-import org.apache.paimon.types.DataTypeCasts;
-import org.apache.paimon.utils.Preconditions;
 
-import org.apache.flink.api.dag.Transformation;
-import org.apache.flink.streaming.api.datastream.DataStream;
-import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
-import org.apache.flink.table.api.EnvironmentSettings;
-import org.apache.flink.table.api.bridge.java.StreamTableEnvironment;
-import org.apache.flink.table.data.RowData;
-import org.apache.flink.table.types.logical.LogicalType;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
-/** Abstract base of {@link Action}. */
+/** Abstract base of {@link Action} for table. */
 public abstract class ActionBase implements Action {
 
-    private static final Logger LOG = LoggerFactory.getLogger(ActionBase.class);
-
-    protected Catalog catalog;
+    protected final Catalog catalog;
 
     protected final FlinkCatalog flinkCatalog;
 
     protected final String catalogName = "paimon-" + UUID.randomUUID();
 
-    protected StreamExecutionEnvironment env;
-
-    protected StreamTableEnvironment tEnv;
-
-    protected Identifier identifier;
-
-    protected Table table;
-
-    ActionBase(String warehouse, String databaseName, String tableName) {
-        this(warehouse, databaseName, tableName, new Options());
+    public ActionBase(String warehouse, Map<String, String> catalogOptions) {
+        this(warehouse, Options.fromMap(catalogOptions));
     }
 
-    ActionBase(String warehouse, String databaseName, String tableName, Options catalogOptions) {
+    public ActionBase(String warehouse, Options catalogOptions) {
         catalogOptions.set(CatalogOptions.WAREHOUSE, warehouse);
-        identifier = new Identifier(databaseName, tableName);
         catalog = FlinkCatalogFactory.createPaimonCatalog(catalogOptions);
         flinkCatalog = FlinkCatalogFactory.createCatalog(catalogName, catalog);
-
-        env = StreamExecutionEnvironment.getExecutionEnvironment();
-        tEnv = StreamTableEnvironment.create(env, EnvironmentSettings.inBatchMode());
-
-        // register flink catalog to table environment
-        tEnv.registerCatalog(flinkCatalog.getName(), flinkCatalog);
-        tEnv.useCatalog(flinkCatalog.getName());
-
-        try {
-            table = catalog.getTable(identifier);
-        } catch (Catalog.TableNotExistException e) {
-            LOG.error("Table doesn't exist in given path.", e);
-            System.err.println("Table doesn't exist in given path.");
-            throw new RuntimeException(e);
-        }
-    }
-
-    /**
-     * Extract {@link LogicalType}s from Flink {@link org.apache.flink.table.types.DataType}s and
-     * convert to Paimon {@link DataType}s.
-     */
-    protected List<DataType> toPaimonTypes(
-            List<org.apache.flink.table.types.DataType> flinkDataTypes) {
-        return flinkDataTypes.stream()
-                .map(org.apache.flink.table.types.DataType::getLogicalType)
-                .map(LogicalTypeConversion::toDataType)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * Check whether each {@link DataType} of actualTypes is compatible with that of expectedTypes
-     * respectively.
-     */
-    protected boolean compatibleCheck(List<DataType> actualTypes, List<DataType> expectedTypes) {
-        if (actualTypes.size() != expectedTypes.size()) {
-            return false;
-        }
-
-        for (int i = 0; i < actualTypes.size(); i++) {
-            if (!DataTypeCasts.supportsCompatibleCast(actualTypes.get(i), expectedTypes.get(i))) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    /** Sink {@link DataStream} dataStream to table with Flink Table API in batch environment. */
-    protected void batchSink(DataStream<RowData> dataStream) {
-        List<Transformation<?>> transformations =
-                Collections.singletonList(
-                        new FlinkSinkBuilder((FileStoreTable) table)
-                                .withInput(dataStream)
-                                .withLockFactory(
-                                        Lock.factory(
-                                                catalog.lockFactory().orElse(null), identifier))
-                                .build()
-                                .getTransformation());
-
-        List<String> sinkIdentifierNames = Collections.singletonList(identifier.getFullName());
-
-        TableEnvironmentUtils.executeInternal(tEnv, transformations, sinkIdentifierNames);
-    }
-
-    /**
-     * The {@link CoreOptions.MergeEngine}s will process -U/-D records in different ways, but we
-     * want these records to be sunk directly. This method is a workaround. Actions that may produce
-     * -U/-D records can call this to disable merge engine settings and force compaction.
-     */
-    protected void changeIgnoreMergeEngine() {
-        if (CoreOptions.fromMap(table.options()).mergeEngine()
-                != CoreOptions.MergeEngine.DEDUPLICATE) {
-            Map<String, String> dynamicOptions = new HashMap<>();
-            dynamicOptions.put(
-                    CoreOptions.MERGE_ENGINE.key(), CoreOptions.MergeEngine.DEDUPLICATE.toString());
-            // force compaction
-            dynamicOptions.put(CoreOptions.FULL_COMPACTION_DELTA_COMMITS.key(), "1");
-            Preconditions.checkArgument(
-                    table instanceof FileStoreTable, "Only supports FileStoreTable.");
-            table = ((FileStoreTable) table).internalCopyWithoutCheck(dynamicOptions);
-        }
     }
 }
