+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/action/CompactAction.java
@@ -46,7 +46,7 @@ import static org.apache.paimon.flink.action.Action.getPartitions;
 import static org.apache.paimon.flink.action.Action.getTablePath;
 
 /** Table compact action for Flink. */
-public class CompactAction extends ActionBase {
+public class CompactAction extends TableActionBase {
 
     private static final Logger LOG = LoggerFactory.getLogger(CompactAction.class);
